exchange_core/requirements.txt
fastapi
uvicorn
pydantic
websockets
psycopg2-binary